requirements.txt
Django==1.11.1
djangorestframework==3.6.3
pymongo==3.4.0
pytz==2017.2
wheel==0.24.0
